Migrate contest#show test to NokoGiri"
